requirements.txt
airtest==1.1.3
certifi==2020.4.5.1
chardet==3.0.4
comtypes==1.1.7
ddt==1.4.1
facebook-wda==0.2.1
hrpc==1.0.8
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
mss==4.0.3
numpy==1.18.4
opencv-contrib-python==3.4.2.17
Pillow==7.1.2
pocoui==1.0.79
PyMySQL==0.9.3
pypiwin32==223
pywin32==227
pywinauto==0.6.3
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.9
websocket-client==0.48.0
